moved and updated scripts with type hints and docstrings
